requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
drf-spectacular==0.27.0
python-dotenv==1.0.0